+++ sqlite/src/commonTest/kotlin/pw/binom/db/sqlite/SQLiteConnectorTest.kt
@@ -4,61 +4,73 @@ import pw.binom.UUID
 import pw.binom.io.file.File
 import pw.binom.io.use
 import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertFalse
-import kotlin.test.assertTrue
 
 class SQLiteConnectorTest {
 
-    val sql = "CREATE TABLE IF NOT EXISTS COMPANY(ID INT PRIMARY KEY NOT NULL, NAME TEXT, uid blob not null)"
+    val sql = "CREATE TABLE IF NOT EXISTS COMPANY(ID INT PRIMARY KEY NOT NULL, NAME TEXT NOT NULL, uid blob not null)"
+
+    @Test
+    fun test2() {
+        val f = File("file.db")
+        try {
+            SQLiteConnector.openFile(f).use {
+                it.createStatement().use {
+                    it.executeUpdate(sql)
+                }
+
+                repeat(5000) { num ->
+                    it.prepareStatement("insert into company (id,name,uid) values(?,?,?)").use {
+                        it.set(0, num)
+                        val s = UUID.random()
+                        it.set(1, s.toString())
+                        val buf = ByteArray(16)
+                        s.toByteArray(buf)
+                        it.set(2, buf)
+                        it.executeUpdate()
+                    }
+//                    it.commit()
+                }
+            }
+        } finally {
+            f.delete()
+        }
+    }
 
     @Test
     fun test() {
-        val file =File("file.db")
-        val uuid = UUID.random()
+        val f = File("file.db")
         try {
-            SQLiteConnector.openFile(file).use {
+            SQLiteConnector.openFile(f).use {
                 it.createStatement().use {
                     it.executeUpdate(sql)
 
                     it.executeQuery("select * from company").use {
                         while (it.next()) {
-                            println("->>${it.getInt(0)}  \"${it.getString(1)}\"  ${UUID.create(it.getBlob(2))}")
+                            println("->>${it.getInt(0)}  \"${it.getString(1)}\"  ${it.getUUID(2)}")
                         }
                     }
                 }
 
-                it.prepareStatement("insert into company (id, name,uid) values(?,?,?)").use {
-                    it.set(0, 1)
-                    it.set(1, uuid.toString())
-                    it.set(2, uuid)
-                    it.executeUpdate()
-
-                    it.set(0, 2)
-                    it.setNull(1)
-                    it.set(2, uuid)
+                it.prepareStatement("insert into company (name,uid) values(?,?)").use {
+                    val s = UUID.random()
+                    it.set(0, s.toString())
+                    val buf = ByteArray(16)
+                    s.toByteArray(buf)
+                    it.set(1, buf)
                     it.executeUpdate()
                 }
 
-                it.prepareStatement("select * from company order by id").use {
+                it.prepareStatement("select * from company where id=?").use {
+                    it.set(0, 1)
                     it.executeQuery().use {
-                        assertTrue(it.next())
-                        assertEquals(1, it.getInt(0))
-                        assertFalse(it.isNull(1))
-                        assertEquals(uuid.toString(), it.getString(1))
-                        assertEquals(uuid, it.getUUID(2))
-
-
-                        assertTrue(it.next())
-                        assertEquals(2, it.getInt(0))
-                        assertTrue(it.isNull(1))
-                        assertEquals(uuid, it.getUUID(2))
+                        while (it.next()) {
+                            println("--->${it.getInt(0)}")
+                        }
                     }
                 }
             }
         } finally {
-            file.delete()
+            f.delete()
         }
-
     }
 }